- updated WriteData to allow for "pointers" from js to work - updated test page to have cleaner output and use WriteData instead of several calls to WriteByte

+++ src/main.cpp
@@ -53,7 +53,7 @@ int main() {
         0xFB
     };
     
-    processor.WriteData(0, program, sizeof(program) / sizeof(uint8_t));
+    processor.WriteData(0, (int)program, sizeof(program) / sizeof(uint8_t));
 
     processor.Execute();
